requirements.txt
Django
dj-rest-auth
django-debug-toolbar
django-extensions
django-filter
djangorestframework
djangorestframework-jsonapi
Faker
